// src/app/enums/worktime-entry-element.enum.ts
export enum WorktimeEntryElement {
    TASK = 'task',
    TIME = 'time',
    STATE_UPDATE_DATE = 'stateUpdateDate',
    HANDLE_TIMER = 'handleTimer',
    RESET_TIMER = 'resetTimer',
    OPEN_TASK = 'openTask',
    MOVE_TO_HISTORY_AND_RESET = 'moveToHistoryAndReset',
    REMOVE_WORKTIME = 'removeWorktime'
}